Implemented dynamic graph re-build
Doing this by a workaround to save and restore the model, as copying from an old to a new graph had problems because all variables from functions that are created with tf.Variable still excist (a lot of reuse and not-initialized errors).

Additionally, we provide options to enable tracking/restoring of moving average variables.