Implemented MsPacman dataset.
The current implementation is quite slow, as it results in many IO operations. Generating a batch of size 32 takes 0.25 - 0.75 sec.
Transforming the Training-set into an input-queue could solve the problem. But research has to be done first on how to fill shuffled batches from multiple files at one. Otherwise, we have to do a very long pre-generation phase to generate up to 5.000.000 small crops, which is not ideal as well.