Added eval_mode param for graph construction
There has been a lot of issues when graph construction, due to LOG_LOSSES that could not be restored. In more detail, the source has been the EMA values of the losses, which are NOT even needed in evaluation mode. This is more kind of a workaround to make the model more robust in evalution, even when the loss-function of the model has changed (e.g new metric or loss func used, triplet losses, etc.)